+++ CHANGELOG.md
@@ -87,3 +87,7 @@ The theme today is getting stateful toggling supported with a number of function
 2021-04-16:
 
 - Add a mutiplier to fakesun() to reduce the sleep interval on the iterations through its loop for increasing brightness. It's a really squishy number; the delay is variable based on the number of lights, the length of the run, whatever processing the bridge has to do, how quickly the lights respond, etc. etc. etc. For my 7-minute cycle with 9 lights, I was actually seeing fakesun() take as much as 14 minutes. This helps a bit. Reducing sleepInterval further runs up against hard limits/race conditions, so.
+
+2021-05-04:
+
+- Moved the light (#35) from the end-table to the coffee station. Update scenes accordingly...